+++ source/Nuke.Components.GitLab/GitLabTasks.cs
@@ -17,7 +17,7 @@ public static class GitLabTasks
 
     public static void Reauthenticate(ICreateGitLabRelease build)
     {
-        ApiClient = new((build.GitLabHostName ?? GitLabHost.Default).EnsureStarting("https://"),
+        ApiClient = new((build.GitLabHostName ?? GitLabHost.Default).EnsureStarting(build.AccessOverHttps ? "https://" : "http://"),
             build.GitLabToken ?? CI.GitLab.GitLab.Instance.JobToken);
     }