+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [vNext]
+- fix: Allow `http://` URLs for GitLab & Forgejo; requires opt-in.
 
 ## [10.3.1] / 2026-04-08
 - Exposed JSON payload in ForgejoActions CI class.